Moved bit_flag from utils to containers
