+++ modules/containers/src/bit_flag.cpp
@@ -2,4 +2,4 @@
 // Copyright (C) 2023-2024 HEPHAESTUS Contributors
 //=================================================================================================
 
-#include "hephaestus/utils/bit_flag.h"  // NOLINT(misc-include-cleaner)
+#include "hephaestus/containers/bit_flag.h"  // NOLINT(misc-include-cleaner)
