+++ modules/utils/CMakeLists.txt
@@ -24,7 +24,6 @@ set(SOURCES
     src/string/string_literal.cpp
     src/string/string_utils.cpp
     src/timing/watchdog.cpp
-    src/bit_flag.cpp
     src/concepts.cpp
     src/exception.cpp
     src/utils.cpp
@@ -38,7 +37,6 @@ set(SOURCES
     include/hephaestus/utils/string/string_literal.h
     include/hephaestus/utils/string/string_utils.h
     include/hephaestus/utils/timing/watchdog.h
-    include/hephaestus/utils/bit_flag.h
     include/hephaestus/utils/concepts.h
     include/hephaestus/utils/exception.h
     include/hephaestus/utils/utils.h
