+++ modules/utils/tests/CMakeLists.txt
@@ -3,13 +3,6 @@
 # =================================================================================================
 
 define_module_test(
-  NAME bit_flag_tests
-  SOURCES bit_flag_tests.cpp
-  PUBLIC_INCLUDE_PATHS $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../include>
-  PUBLIC_LINK_LIBS ""
-)
-
-define_module_test(
   NAME concepts_tests
   SOURCES concepts_tests.cpp
   PUBLIC_INCLUDE_PATHS $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/../include>
